+++ packages/expo-router/src/react-navigation/material-top-tabs/__tests__/missing-dependency.test.ios.tsx
@@ -9,6 +9,9 @@ jest.mock(
 );
 
 test('throws an error when react-native-tab-view is not installed', () => {
+  jest.doMock('react-native-tab-view', () => {
+    throw new Error();
+  });
   expect(() => require('../index')).toThrow(
     "Install the 'react-native-tab-view' package and its peer dependencies to use the Expo Router's TopTabs."
   );